Use "this" instead of "wordbreak" to make wordbreak extendable
We want to extend wordbreak.isBreak method, because apostrophe is not
sometimes an wordbreak in different languages. So when we extend it,
wordbreak.prevBreakOffset method should use the new extended version of
isBreak method. But it always uses the wordbreak.isBreak method, that's
why the extended isBreak method is never used. If the prevBreakOffset
and other methods uses this.isBreak, we can easily extend the isBreak
method.

I could not run the unit tests even in master branch. But I cloned the
unicodejs and tested the approach I suggested, it worked well in our
product.

Bug: T270706
Change-Id: Ia40d1b562017c6c4283b6af62de1f180dc1e0b90